feat: implementada request/response com XML para o Account

+++ app/Services/AccountService.php
@@ -6,9 +6,7 @@ use App\Repositories\Account\AccountRepositoryInterface;
 use App\Repositories\Account\AccountTypeRepositoryInterface;
 use App\Repositories\Customer\CustomerRepositoryInterface;
 
-/**
- * Summary of AccountService
- */
+
 class AccountService
 {
     private $repoAccount, $repoAccountType, $repoCustomer;
@@ -30,24 +28,25 @@ class AccountService
     /**
      * Envia para o AccountRepository os dados para criar uma nova instância de Account
      * @param \Illuminate\Support\Collection|array|int|string $data
-     * @return \Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response
+     * @return array
      */
     public function store($data)
     {
         $errors = [];
 
         try {
+            // TODO: Tratar os dados da requisição, antes de chamar o repoAccount->store
             $currentCustomer = $this->repoCustomer->get($data['fk_customer']);
             $currentAccountType = $this->repoAccountType->get($data['fk_account_type']);
 
-            return response(
-                json_encode([
+            return [
+                'data' => [
                     'account' => $this->repoAccount->store($data),
                     'customer' => $currentCustomer,
                     'account_type' => $currentAccountType,
-                ]),
-                201
-            );
+                ],
+                'status' => 201
+            ];
 
         } catch (\Throwable) {
             if ($currentCustomer == null)
@@ -55,12 +54,10 @@ class AccountService
             if ($currentAccountType == null)
                 array_push($errors, 'Account type not found');
 
-            return response(
-                json_encode([
-                    'msg' => $errors,
-                ]),
-                404
-            );
+            return [
+                'message' => $errors,
+                'status' => 404
+            ];
         }
     }
 
@@ -76,31 +73,86 @@ class AccountService
     /**
      * Retorna uma instância de Account a partir do id informado
      * @param int|string $id
-     * @return \App\Models\Account
+     * @return array
      */
     public function get($id)
     {
-        return $this->repoAccount->get($id);
+        try {
+            $account = $this->repoAccount->get($id);
+            $account->accountType;
+            $account->customer;
+            $status = 200;
+
+            throw_if($account == null);
+
+            return compact('account', 'status');
+
+        } catch (\Throwable) {
+            return [
+                'message' => 'Account not found',
+                'status' => 404
+            ];
+        }
     }
 
     /**
      * Atualiza os dados de uma instância de Account
      * @param \Illuminate\Support\Collection|array|int|string $data
      * @param int|string $id
-     * @return \App\Models\Account
+     * @return array
      */
     public function update($data, $id)
     {
-        return $this->repoAccount->update($data, $id);
+        try {
+            // TODO: Tratar os dados da requisição, antes de chamar o repoAccount->update
+            $keys = [];
+            $values = [];
+            foreach ($data as $key => $value) {
+                array_push($keys, $key);
+                array_push($values, $value);
+            }
+
+            $processed_data = array_combine($keys, $values);
+
+            $status = 200;
+
+            $this->repoAccount->update($processed_data, $id);
+            $account = $this->repoAccount->get($id);
+            $status = 200;
+
+            throw_if($account == null);
+
+            return compact('account', 'status');
+
+        } catch (\Throwable) {
+            return [
+                'message' => 'Account not found',
+                'status' => 404
+            ];
+        }
     }
 
     /**
      * Remove uma instância de Account do banco de dados
      * @param int|string $id
-     * @return int
+     * @return array
      */
     public function destroy($id)
     {
-        return $this->repoAccount->destroy($id);
+        $account = $this->repoAccount->get($id);
+
+        if ($account == null) {
+            $message = 'Account not found';
+            $status = 404;
+        } else if (count($account->transactions) > 0) {
+            $message = 'This account has associated transactions';
+            $status = 405;
+        } else {
+            $this->repoAccount->destroy($id);
+            $message = 'Account destroyed';
+            $status = 204;
+        }
+
+        return compact('message', 'status');
     }
 }
